fix(s05): address Copilot review on PR #101
Six Copilot comments on PR #101 all legitimate and addressed:

1. timescale_password default SecretStr("apex_secret") removed —
   defaults to SecretStr(""), env var required. Aligns with docstring
   that says "sensitive values must be set via environment".

2-6. Mixed Decimal/float comparisons in S05 risk thresholds:
   - MIN_RR_RATIO, MAX_DAILY_LOSS_PCT, MAX_INTRADAY_LOSS_30M_PCT,
     MAX_TOTAL_EXPOSURE_PCT, MAX_CLASS_EXPOSURE_PCT, MAX_RISK_PER_TRADE_PCT,
     MAX_POSITION_SIZE_PCT, CRYPTO_SIZE_MULTIPLIER, CB_SCALP_SIZE_MULTIPLIER
     converted to Decimal constants in services/s05_risk_manager/models.py.
   - VIX_SPIKE_THRESHOLD_PCT, MIN_META_CONFIDENCE, MIN_KELLY_FRACTION,
     CORRELATION_BLOCK_THRESHOLD kept as float (compared with float values).
   - Removed redundant Decimal(str(CONST)) wrappers in position_rules.py
     and service.py — constants are now Decimal at source.
   - cb_event_guard.py: get_post_event_size_multiplier() return type
     updated float → Decimal.

Rationale: risk-critical thresholds must be exact. Binary float
representation of 0.03 is 0.029999... which subtly shifts CB trip points.

Quality gates: ruff clean, mypy strict 0 errors, 1228 tests passing.

Refs #66 #71, PR #101

+++ services/s05_risk_manager/cb_event_guard.py
@@ -26,6 +26,7 @@ from __future__ import annotations
 
 import json
 from datetime import UTC, datetime, timedelta
+from decimal import Decimal
 
 import structlog
 from redis.asyncio import Redis
@@ -111,7 +112,7 @@ class CBEventGuard:
         return False
 
     @staticmethod
-    def get_post_event_size_multiplier() -> float:
+    def get_post_event_size_multiplier() -> Decimal:
         """Return the size multiplier applied during post-event scalp windows.
 
         Returns:
